frontend: move route state to api service
Will be reused a lot, so may as well.
